Move custom validations to gem-level validation files to DRY
