+++ app/models/newspaper_article.rb
@@ -12,23 +12,12 @@ class NewspaperArticle < ActiveFedora::Base
   # containment/aggregation:
   self.valid_child_concerns = [NewspaperPage]
 
+  validates_with NewspaperWorks::PublicationDateValidator
+
   # Validation and required fields:
   validates :title, presence: {
     message: 'A newspaper article requires a title.'
   }
-
-  validate :publication_date_valid
-
-  def publication_date_valid
-    error_msg = "Incorrect Date. Date input should be formatted yyyy-mm-dd and be a valid date."
-    return unless publication_date.present?
-    unless DATE_REGEX.match(publication_date)
-      errors.add(:publication_date, error_msg)
-      return
-    end
-    date_split = publication_date.split("-").map(&:to_i)
-    errors.add(:publication_date, error_msg) unless Date.valid_date?(date_split[0], date_split[1], date_split[2])
-  end
 
   # TODO: Implement validations
   # validates :resource_type, presence: {
