+++ app/models/newspaper_container.rb
@@ -16,30 +16,7 @@ class NewspaperContainer < ActiveFedora::Base
     message: 'A newspaper container requires a title.'
   }
 
-  validate :publication_date_start_valid,
-           :publication_date_end_valid,
-           :publication_date_start_before_publication_date_end
-
-  @date_incorrect_error_msg = "Incorrect Date. Date input should be formatted yyyy[-mm][-dd] and be a valid date."
-  def publication_date_start_valid
-    return if !publication_date_start.present? || publication_date_valid?(publication_date_start)
-    errors.add(:publication_date_start, @date_incorrect_error_msg)
-  end
-
-  def publication_date_end_valid
-    return if !publication_date_end.present? || publication_date_valid?(publication_date_end)
-    errors.add(:publication_date_end, @date_incorrect_error_msg)
-  end
-
-  def publication_date_start_before_publication_date_end
-    return unless publication_date_start.present? && publication_date_end.present?
-    pub_start = publication_date_start.split("-")
-    pub_end = publication_date_end.split("-")
-    (0..2).each do |i|
-      errors.add(:publication_date_start, "Publication start date must be earlier or the same as end date.") if pub_start[i] && pub_end[i] && pub_end[i] < pub_start[i]
-      break
-    end
-  end
+  validates_with NewspaperWorks::PublicationDateStartEndValidator
 
   # TODO: Implement validations
   # validates :resource_type, presence: {
@@ -106,14 +83,4 @@ class NewspaperContainer < ActiveFedora::Base
   def pages
     members.select { |v| v.instance_of?(NewspaperPage) }
   end
-
-  private
-
-    def publication_date_valid?(pub_date)
-      return false unless DATE_RANGE_REGEX.match(pub_date)
-      date_split = pub_date.split("-").map(&:to_i)
-      return false if date_split.length == 3 &&
-                      !Date.valid_date?(date_split[0], date_split[1], date_split[2])
-      true
-    end
 end
